Player.h
So.... I added a bunch of stuff to player.  I think the biggest
difficulty is that I'm not entirely sure if I'm asking for the correct
input from GameMao.  I'd like to ask for the player vector that holds
0-3 and an actual number that represents which player 1-4 to be handed
to getName.  Check it out and let me know if there is anything "off".