Changed shuffle, deal, and playerPlay a bit...
Shuffle needed a check to see if discard actually had cards in it,
otherwise we would get a divide by zero.

Put srand() in deal to randomize things (because deal only run once).

Changed int in deal's for loop to unsigned int because players.size()
returns unsigned and can only compare those.

Got rid of an extra card that was being placed on the discard.

Also won't allow the player to draw from an empty deck. (In playerPlay)